Updated HTML and CSS for forms
This commit is the updated HTML and CSS files for my TrackTour app.  The
html has been modified to accommodate a separate space for people to add
artists to the application.  Additionally, there is a separate form
created that allows for users to add artists to their form.  I also
added the table where the user's list of artist's will be rendered.